[#576] MessageHelper improvements.
